added extended slice in auth
