+++ frontend/src/app/api/apiSlice.js
@@ -1,0 +1,19 @@
+import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react";
+
+const baseQuery = fetchBaseQuery({
+	baseUrl: "http://127.0.0.1:8000",
+	prepareHeaders: (headers, { getState }) => {
+		const token = getState().auth.token;
+		if (token) {
+			headers.set("Authorization", `Token ${token}`);
+		}
+		return headers;
+	},
+});
+
+export const apiSlice  = createApi({
+	baseQuery: baseQuery,
+	endpoints: () => ({}),
+});
+
+
