+++ frontend/src/features/auth/index.js
@@ -1,4 +1,6 @@
 import Auth from "features/auth/Auth";
 import authSlice from "features/auth/authSlice";
+import {useLoginUserMutation} from "features/auth/authApiSlice";
 
-export { Auth, authSlice };
+
+export  { Auth, authSlice, useLoginUserMutation };
